% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_options.R
\name{assert_fxoptions}
\alias{assert_fxoptions}
\title{Verify that x is a valid fx_options object}
\usage{
assert_fxoptions(x)
}
\arguments{
\item{x}{An object to verify}
}
\value{
Unchanged input if valid, otherwise an error is thrown.
}
\description{
Verifies that \code{x} is an \code{fx_options} object (of class
\code{fxconvert_fx_options}) and that all elements of the object are
valid for such an object. Use \code{fx_options()} to create \code{fx_options}
objects.
}
\keyword{internal}
