% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_imports_and_functions.R
\name{assert_nodots}
\alias{assert_nodots}
\title{Assert that no dots arguments are passed}
\usage{
assert_nodots(...)
}
\arguments{
\item{...}{The \code{ellipsis} argument should be empty.}
}
\description{
Asserts that no named or unnamed parameters to the
function end up assigned to the \code{ellipsis} or \code{dots} argument.
The function throws an error if any parameters match the
argument.
}
\keyword{internal}
