% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{double_date}
\alias{double_date}
\title{Convert an Object to a Double-Preserved Date}
\usage{
double_date(x)
}
\arguments{
\item{x}{An object that can be coerced into a date-like numeric format.}
}
\value{
A \code{Date} object stored as a \code{double}.
}
\description{
This function converts an object to a numeric double representation
while preserving the "Date" class. It first unclasses the input,
converts it to \code{double}, and then reassigns the \code{"Date"} class.
}
\details{
Used to ensure that dates read using \code{nanoparquet} are identical to
dates read from the same file using \code{arrow}
}
\examples{
  d <- 10957L
  class(d) <- "Date" # d is now "2000-01-01"
  double_date(d)  # Returns the same date but stored as a double

}
\keyword{internal}
