% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_date_seq_lumpy.R
\name{fx_date_seq_lumpy}
\alias{fx_date_seq_lumpy}
\title{Generate a lumpy sequene of dates}
\usage{
fx_date_seq_lumpy(from_date, to_date)
}
\arguments{
\item{from_date}{A date or character string in "YYYY-MM-DD" format.}

\item{to_date}{A date or character string in "YYYY-MM-DD" format.}
}
\value{
A character vector of separated date ranges, sorted chronologically.
The vector contains strings representing full years ("YYYY"),
full months ("YYYY-MM"), and individual days ("YYYY-MM-DD") as needed.
}
\description{
Generates a "lumpy" sequence of dates in that the dates are grouped into
individual components of full years, full months, and remaining days.
}
\examples{
fx_date_seq_lumpy("2022-11-30", "2024-02-01")
}
\keyword{internal}
