% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_get.R
\name{fx_get}
\alias{fx_get}
\alias{fx_get_single}
\title{Retrieve Foreign Exchange Rates for a Specific Date}
\usage{
fx_get(from, to, fxdate, fxsource = "ecb", ..., .interpolate = FALSE)

fx_get_single(from, to, fxdate, fxsource = "ecb", ..., .interpolate = FALSE)
}
\arguments{
\item{from}{The ISO currency code (as a character string) for the base currency.}

\item{to}{The ISO currency code (as a character string) for the target currency.}

\item{fxdate}{The date for which to retrieve exchange rates, in "YYYY-MM-DD" format.}

\item{fxsource}{A character string specifying the source of the exchange rate data.
Default is "ecb" (European Central Bank).}

\item{...}{Reserved}

\item{.interpolate}{If a weekend date is requested, should the previous days rate
be returned. If this is false, a weekend date will result in an error.}
}
\value{
A numeric value representing the exchange rate from the \code{from} currency
to the \code{to} currency on the specified \code{fxdate}. The exchange rate
between\code{from} and \code{to}, given in a way that to convert, from \code{from}
to \code{to} one would multiply by the exchange rate. This means that the
function may not always give the expected value for well-known
currency pairs that are always quoted in the same direction.
}
\description{
This function returns the exchange rates between two specified currencies for
a given date, and calculates the exchange rate from the first specified
currency to the second.
}
\examples{
#fx_get("USD", "EUR", "2022-01-01")

}
