% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fx_get_fxdata_dir}
\alias{fx_get_fxdata_dir}
\title{Get Local FX Data Directory}
\usage{
fx_get_fxdata_dir(..., options = fx_options())
}
\arguments{
\item{...}{Reserved. Options parameter must be named}

\item{options}{A valid \code{fx_options} object.}
}
\value{
A character string representing the path to the local FX data
directory.
}
\description{
Returns the path to the local directory for storing FX data.
}
\keyword{internal}
