% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_get.R
\name{fx_get_single}
\alias{fx_get_single}
\title{This function connects to a DuckDB database containing foreign exchange rate data,
retrieves exchange rates between two specified currencies for a given date,
and calculates the exchange rate from the first specified currency to the second.}
\usage{
fx_get_single(from, to, fxdate, bank = "ecb", ..., .interpolate = FALSE)
}
\description{
This function connects to a DuckDB database containing foreign exchange rate data,
retrieves exchange rates between two specified currencies for a given date,
and calculates the exchange rate from the first specified currency to the second.
}
\keyword{internal}
