% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_init.R
\name{fx_init}
\alias{fx_init}
\title{Initialize Foreign Exchange (FX) Data}
\usage{
fx_init(
  ...,
  banks = c("ecb", "cbi", "fed", "xfed"),
  action = c("auto", "update", "offline", "full", "remove"),
  verbose = TRUE,
  once = FALSE,
  mirai = FALSE
)
}
\arguments{
\item{...}{Reserved. All arguments must be named.}

\item{banks}{Character string specifying the source of FX data. Currently,
\code{"ecb"}, \code{"cbi"}, and \code{"fed"} are supported.}

\item{action}{String specifying which initialization action to take. One of
\code{auto}, \code{update}, \code{offline}, \code{full}, \code{remove}. See details. Replaces
\code{approach}.}

\item{verbose}{Logical indicating whether to print progress messages.
Defaults to \code{TRUE}.}

\item{once}{Logical indicating whether to only perform actual initialization
once per R session. Defaults to \code{FALSE}.}

\item{mirai}{Should data fetching use \code{mirai} for parallel processing.}
}
\description{
This function initializes and refreshes the local foreign exchange (FX) data
directory by downloading the necessary data files from the specified remote
server. The function allows different approaches for data management,
including incremental updates, full re-initialization, and local refreshes.
}
\details{
This function ensures that the local FX data store is up to date by:
\itemize{
\item Checking available data ranges on the remote server.
\item Identifying missing data files for download.
\item Removing outdated local files.
\item Loading data into a DuckDB database for efficient querying.
}

The approach to updating data is determined by the \code{approach} parameter. This
can be currently set to four different options:
\itemize{
\item \code{"incremental"}: Downloads missing data while keeping existing files.
\item \code{"fresh"}: Deletes all local data and downloads everything anew.
\item \code{"local_refresh"}: Rebuilds the local database without downloading new data.
\item \code{"remove"}: Deletes all local data and exits without downloading new data.
}

The different approaches, or actions, are subject to change. In particular
they should support offline use with appropriate warnings. Potential actions:
\itemize{
\item \code{auto}    (equivalent to \code{once}=TRUE, only updates if it has not been updated in this session (but must ensure sitrep is OK))
\item \code{update}  (equivalent to older \code{incremental}, warns on offline, error if no data)
\item \code{offline} (avoid internet access, no warn on offline unless data is missing)
\item \code{full}    (remove and update rolled into one)
\item \code{remove}  (deletes all local data and exits without downloading new data.)
}

TODO: This should take an options object and also
}
