% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_date_seq_lumpy.R
\name{fx_lump_dates}
\alias{fx_lump_dates}
\title{Generate a lumpy sequence of dates with a recursive approach}
\usage{
fx_lump_dates(dates, lump_from = "millennium", lump_to = "decaday")
}
\arguments{
\item{dates}{Strictly increasing date sequence to lump}

\item{lump_from}{Largest era unit to lump}

\item{lump_to}{Smallest era unit to lump}
}
\value{
A character vector of date ranges.
}
\description{
The function lumps dates into common eras, with eras of different length,
ranging from \code{millennium} to \code{decaday}. A \code{decaday} is the set of days in a
month that have a common number in the tens-place (1-9, 10-19, 20-29, 30-31).

Allowed lump eras are: \code{millennium}, \code{century}, \code{decade}, \code{year}, \code{month},
and \code{decaday}.

It calls a non-exported function to do the actual recursion, after doing some
of the more expensive input checking only on the initial input.
}
\keyword{internal}
