% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_options.R
\name{fx_options}
\alias{fx_options}
\title{Create an fx_options Object}
\usage{
fx_options(
  ...,
  workspace = getOption("fxconvert.workspace", "main"),
  bank = getOption("fxconvert.bank", "ecb"),
  repo = getOption("fxconvert.repo", "torfason/fxdata")
)
}
\arguments{
\item{...}{Reserved. All arguments must be named.}

\item{workspace}{A character string specifying the local workspace to use
for operations.}

\item{bank}{A character string specifying which central bank rates to
base the conversions on.}

\item{repo}{A character string specifying the remote repository source for
exchange rate data.}
}
\value{
A list representing the \code{fx_options} object.
}
\description{
Constructs an \code{fx_options} object with configurable parameters.
The parameters can be set from the defaults, by
}
