% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_sitrep.R
\name{fx_sitrep}
\alias{fx_sitrep}
\title{Print debug information about the exchange rate database}
\usage{
fx_sitrep(fxsource = "ecb")
}
\arguments{
\item{fxsource}{Character string specifying the source of exchange rate data.
Defaults to "ecb". Must match a valid source.}
}
\description{
Provides a status report for the exchange rate database, printing key
information about the database file and available data, including
a preview of available dates and consistency checks.
}
\keyword{internal}
