% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fx_vec_fill_gaps}
\alias{fx_vec_fill_gaps}
\title{Fill Missing Values Within Observed Data Range}
\usage{
fx_vec_fill_gaps(
  x,
  direction = c("down", "up", "downup", "updown"),
  max_fill = NULL
)
}
\arguments{
\item{x}{An atomic vector containing \code{NA} values to be filled.}

\item{direction}{A string indicating the fill direction. Must be one of:
\itemize{
\item \code{"down"} (default): Fill missing values downward.
\item \code{"up"}: Fill missing values upward.
\item \code{"downup"}: Fill first down, then up.
\item \code{"updown"}: Fill first up, then down.
}}

\item{max_fill}{A single positive integer specifying the maximum number of
sequential missing values that will be filled. If NULL, there is no limit.}
}
\value{
A vector of the same type as \code{x}, with missing values filled within
the index range of observed (non-missing) values
}
\description{
Fills \code{NA} values in a vector within the index range of observed
(non-missing) values, without extending beyond the last non-missing value.
}
\details{
The function identifies the first and last non-missing values in \code{x} and only
fills missing values within this index range. Any values beyond the last
non-missing value remain \code{NA}.

This function is useful when you need to fill missing values but want to
avoid extending beyond known data points. Internally it
uses \code{\link[vctrs:vec_fill_missing]{vctrs::vec_fill_missing()}} to do the filling.

The \code{"downup"} and \code{"updown"} approaches are provided strictly for
compatibility with \code{\link[vctrs:vec_fill_missing]{vctrs::vec_fill_missing()}}. In practice, the result will
always be exactly equal to just using \code{"up"} or \code{"down"}, respectively.
}
\examples{
x1 <- c(NA, NA, 1, NA, NA, 2, NA, NA)
fx_vec_fill_gaps(x1)
fx_vec_fill_gaps(x1, direction = "up")

x2 <- c(NA, NA, 1, NA, NA)
fx_vec_fill_gaps(x2)  # Unchanged (single element means no gaps)

x3 <- c(NA, NA, NA, NA, NA)
fx_vec_fill_gaps(x3)  # Unchanged (all NA)

}
\keyword{internal}
