% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fx_verify_data_version}
\alias{fx_verify_data_version}
\title{Verify and clean FX data versioning}
\usage{
fx_verify_data_version(d)
}
\arguments{
\item{d}{A data frame with a \code{.version.} column.}
}
\value{
The input data frame with the \code{.version.} column removed.
}
\description{
Checks the \code{.version.} column of an FX dataset. Issues a warning if data is
unversioned or contains multiple versions. Removes the \code{.version.} column
before returning the data.
}
\keyword{internal}
