% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a1_imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ymd}
\alias{today}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lubridate}{\code{\link[lubridate:today]{today()}}, \code{\link[lubridate:ymd]{ymd()}}}
}}

