% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_parquet_vx}
\alias{write_parquet_vx}
\title{Writes and verifies parquet file with maximum compression}
\usage{
write_parquet_vx(
  x,
  file,
  ...,
  compression_types = c("gzip", "snappy", "uncompressed"),
  verbose = FALSE
)
}
\arguments{
\item{x}{\code{tibble} with data to write to file.}

\item{file}{\code{string} with path name to write.}

\item{compression_types}{\code{character} with list of compression methods to try.}

\item{verbose}{\code{flag} determining verbosity level}
}
\value{
\code{TRUE} if data was written to file, \code{FALSE} if a file existed that
contained the exact same data as \code{x} (if a file with different data was
found, an error is thrown).
}
\description{
For ma\strong{X}imum compression, the function writes \code{length(compressions)} versions
of parquet file with maximum compression options for each compression type to
a temporary location, before moving the maximally compressed file to the
target file.

For \strong{V}erification, the function checks if the target file exists, if it
does it is read to verify that its contents equal \code{x}, writing nothing and
returning \code{FALSE} if they are equal (a mismatch is an error). If the target
file does not exist, it is written (see above) and \code{TRUE} returned , then
re-read to ensure a match (again, a mismatch is an error).
}
\keyword{internal}
