% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_convert.R, R/fx_get.R
\name{fx_convert}
\alias{fx_convert}
\alias{fx_get}
\title{Retrieve exchange rates or convert amounts between currencies}
\usage{
fx_convert(amount, from, to, fxdate, bank = "ecb", ..., .interpolate = FALSE)

fx_get(from, to, fxdate, bank = "ecb", ..., .interpolate = FALSE)
}
\arguments{
\item{amount}{Numeric; the amount of money to convert from the \code{from} currency
to the \code{to} currency. Only used in \code{fx_convert()}.}

\item{from}{The ISO currency code (as a character string) for the base currency.}

\item{to}{The ISO currency code (as a character string) for the target currency.}

\item{fxdate}{The date for which to retrieve exchange rates, in "YYYY-MM-DD" format.}

\item{bank}{A character string specifying the source of the exchange rate data.
Default is "ecb" (European Central Bank).}

\item{...}{Reserved}

\item{.interpolate}{If a weekend date is requested, should the previous days rate
be returned. If this is false, a weekend date will result in an error.}
}
\value{
For \code{fx_convert()}, the converted amount in the \code{to} currency.
\code{fx_convert()} relies directly on the exchange rates retrieved using
\code{fx_get()}, and simply multiplies the amount with the rates that \code{fx_get()}
return.

For \code{fx_get()}, a numeric value representing the exchange rate from
the \code{from} currency to the \code{to} currency on the specified \code{fxdate}. The
exchange rate between\code{from} and \code{to}, given in a way that to convert, from
\code{from} to \code{to} one would multiply by the exchange rate. This means that the
function may not always give the expected value for well-known currency
pairs that are always quoted in the same direction.
}
\description{
These functions are used to convert amounts from one currency to another, or
to directly return the exchange rates between two specified currencies.
}
